Estilo y login

+++ app/Http/Controllers/EmpleadoController.php
@@ -48,7 +48,8 @@ class EmpleadoController extends Controller
         }
         //agarra el modelo e inserta los datos
         Empleado::insert($datosEmpleado);
-        return response()->json($datosEmpleado);
+        //return response()->json($datosEmpleado);
+         return redirect('empleado')->with('mensaje','Empleado agregado con exito');
     }
 
     /**
@@ -109,7 +110,11 @@ class EmpleadoController extends Controller
     public function destroy($id)
     {
         //
-        Empleado::destroy($id);
-        return redirect('empleado');
+        $empleado=Empleado::findOrFail($id);
+        if(Storage::delete('public/'.$empleado->Foto)){
+            Empleado::destroy($id);
+        }
+        
+        return redirect('empleado')->with('mensaje','Empleado Borrado');
     }
 }
